Ninja does not quite support this extension arrangement.
- Move `_parseArguments` into UserDefaults.swift.
